+++ Course2FinalTask/Code/Modules/Login/API/Responses/LoginResponse.swift
@@ -1,0 +1,16 @@
+//
+//  LoginResponse.swift
+//  Course2FinalTask
+//
+//  Created by Андрей Груненков on 03.12.2020.
+//  Copyright © 2020 e-Legion. All rights reserved.
+//
+
+import Foundation
+
+struct LoginResponse: Decodable {
+    
+    var token: String
+    
+}
+
