+++ Course2FinalTask/Code/Modules/Users/Views/Extensions/UsersViewController+TableView.swift
@@ -1,0 +1,35 @@
+//
+//  UsersViewController+TableView.swift
+//  Course2FinalTask
+//
+//  Created by Андрей Груненков on 30.04.2021.
+//  Copyright © 2021 e-Legion. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+extension UsersViewController: UITableViewDelegate, UITableViewDataSource {
+    
+    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
+        return 45.0
+    }
+     
+     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return users.count
+     }
+     
+     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+         let cell = usersTable.dequeueReusableCell(withIdentifier: String(describing: UsersCell.self)) as! UsersCell
+         cell.setUser(user: users[indexPath.row])
+         return cell
+     }
+    
+    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        let storyboard = UIStoryboard(name: "Main", bundle: nil)
+        let profileController = storyboard.instantiateViewController(withIdentifier: String(describing: ProfileViewController.self)) as! ProfileViewController
+        profileController.dataSource.user = self.users[indexPath.row]
+        navigationController?.pushViewController(profileController, animated: true)
+    }
+    
+}
