+++ Course2FinalTask/Code/Modules/Users/Views/UsersViewController.swift
@@ -21,36 +21,16 @@ final class UsersViewController: UIViewController {
         setupViews()
     }
     
+}
+
+private extension UsersViewController {
+    
     // MARK: setup
     
-    private func setupViews() {
+    func setupViews() {
         self.usersTable.delegate = self
         self.usersTable.dataSource = self
     }
     
 }
 
-extension UsersViewController: UITableViewDelegate, UITableViewDataSource {
-    
-    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return 45.0
-    }
-     
-     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return users.count
-     }
-     
-     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-         let cell = usersTable.dequeueReusableCell(withIdentifier: String(describing: UsersCell.self)) as! UsersCell
-         cell.setUser(user: users[indexPath.row])
-         return cell
-     }
-    
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        let storyboard = UIStoryboard(name: "Main", bundle: nil)
-        let profileController = storyboard.instantiateViewController(withIdentifier: String(describing: ProfileViewController.self)) as! ProfileViewController
-        profileController.user = self.users[indexPath.row]
-        navigationController?.pushViewController(profileController, animated: true)
-    }
-    
-}
